refactor(builder): split schema generator into cohesive files
Extract the independent top-level declarations from the ~1034-LOC
hand-written schema generator source into cohesive same-library part
files, keeping the tightly-coupled SchemaGenerator core whole.

- schema_generator.collection_info.dart: SchemaCollectionInfo data class
  (the externally-consumed collection-annotation model).
- schema_generator.class_type.dart: ClassType enum plus the
  preferInlineAnnotation / overrideAnnotation emit primitives.

Pure mechanical move via part/part-of. Zero logic change. The public
API surface is byte-for-byte preserved: every top-level name remains
importable identically through the same library, so consumers
(src/schema_generator.dart, builder.dart, and the public export) are
unchanged. The SchemaGenerator class (20 mutually-referencing static
methods plus shared private helpers) cannot be split across Dart part
files and is intentionally kept intact.

Gates vs baseline (dart 3.12.2, firestore_odm_builder package):
- dart analyze: 129 issues, identical to baseline, RC=0 (info-only).
- dart format --set-exit-if-changed: new files clean (0 changed); the
  package-wide changed-file set is unchanged from baseline (pre-existing
  formatter-version skew on untouched files).
