+++ packages/firestore_odm_builder/lib/src/generators/schema_generator.class_type.dart
@@ -1,0 +1,23 @@
+part of 'schema_generator.dart';
+
+final preferInlineAnnotation = refer(
+  'pragma',
+).call([literalString('vm:prefer-inline')]);
+final overrideAnnotation = refer('override');
+
+enum ClassType {
+  root('Root'),
+  collection('Collection'),
+  document('Document'),
+  transactionContext('TransactionContext'),
+  transactionCollection('TransactionCollection'),
+  transactionDocument('TransactionDocument'),
+  batchContext('BatchContext'),
+  batchCollection('BatchCollection'),
+  batchDocument('BatchDocument'),
+  patchBuilder('PatchBuilder');
+
+  const ClassType(this.suffix);
+
+  final String suffix;
+}
